Refactor user csv download into helper
Kept the Response (and mimetype/headers) in the views as they are
more 'Flasky'
